Basic color pallete.
