+++ packages/theme/functions.php
@@ -2482,193 +2482,9 @@ function custom_parent_dropdown_limit($args, $request)
 }
 add_filter('rest_page_query', 'custom_parent_dropdown_limit', 20, 2);
 // theme options
-add_action('customize_register', 'theme_customize_register');
-function theme_customize_register($wp_customize)
-{
-    $wp_customize->add_panel('theme', array(
-        'title'=> 'Ustawienia motywu',
-        'priority' => 160,
-    ));
-    $wp_customize->add_section('main', array(
-        'title'=>'Główne',
-        'panel'=>'theme'
-    ));
-    $wp_customize->add_setting('mourning', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('mourning', array(
-        'type'=>'checkbox',
-        'priority'=>10,
-        'section'=>'main',
-        'label'=>'Żałoba',
-        'description'=>'Cała strona WWW zostaje wyświetlona w odcieniach szarości.'
-    ));
-    $wp_customize->add_setting('add_event', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('add_event', array(
-        'type'=>'checkbox',
-        'priority'=>10,
-        'section'=>'event',
-        'label'=>'Dodawanie wydarzeń',
-        'description'=>'Niezalogowani użytkownicy mogą dodaważ wydarzenia'
-    ));
-    $wp_customize->add_setting('logo', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_setting('header_logo_1', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-
-    $wp_customize->add_setting('header_logo_2', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-
-    $wp_customize->add_setting('header_logo_3', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-
-
-
-    $wp_customize->add_control(new WP_Customize_Media_Control($wp_customize, 'header_logo_1', array(
-        'priority'=>10,
-        'section'=>'main',
-        'label'=>'Nagłówek - logo - 1',
-        'description'=>'Logo #1 w nagłówku strony',
-        'mime_type' => 'image'
-    )));
-    $wp_customize->add_control(new WP_Customize_Media_Control($wp_customize, 'header_logo_2', array(
-        'priority'=>10,
-        'section'=>'main',
-        'label'=>'Nagłówek - logo - 2',
-        'description'=>'Logo #2 w nagłówku strony',
-        'mime_type' => 'image'
-    )));
-    $wp_customize->add_control(new WP_Customize_Media_Control($wp_customize, 'header_logo_3', array(
-        'priority'=>10,
-        'section'=>'main',
-        'label'=>'Nagłówek - logo - 3',
-        'description'=>'Logo #3 w nagłówku strony',
-        'mime_type' => 'image'
-    )));
-    $wp_customize->add_control(new WP_Customize_Media_Control($wp_customize, 'logo', array(
-        'priority'=>10,
-        'section'=>'main',
-        'label'=>'Logo',
-        'description'=>'Logo zostanie wyświetlone w nagłówku strony.',
-        'mime_type' => 'image'
-    )));
-    $wp_customize->add_setting('placeholder', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control(new WP_Customize_Media_Control($wp_customize, 'placeholder', array(
-        'priority'=>10,
-        'section'=>'main',
-        'label'=>'Placeholder',
-        'description'=>'Placeholder zostanie wyświetlone w momencie gdy nie ustawisz zdjęcia wyróżnionego wpisu.',
-        'mime_type' => 'image'
-    )));
-    $wp_customize->add_section('event', array(
-        'title'=>'Wydarzenia',
-        'panel'=>'theme',
-    ));
-    $wp_customize->add_setting('domains', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('domains', array(
-        'type'=>'input',
-        'priority'=>10,
-        'section'=>'event',
-        'label'=>'Dozwolone domeny',
-        'description'=>'Wprowadź po przecinku domeny dla maili z których można zgłaszać wydarzenia.'
-    ));
-    $wp_customize->add_setting('subject', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('subject', array(
-        'type'=>'input',
-        'priority'=>10,
-        'section'=>'event',
-        'label'=>'Tytuł maila z linkiem do potwierdzenia',
-        'description'=>'Wprowadź tytuł maila który zostanie wysłany z tokenem do weryfikacji.'
-    ));
-    $wp_customize->add_setting('message', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('message', array(
-        'type'=>'textarea',
-        'priority'=>10,
-        'section'=>'event',
-        'label'=>'Treść maila z linkiem do potwierdzenia',
-        'description'=>'Wprowadź treść maila który zostanie wysłany z tokenem do weryfikacji. Użyj {{ token }} w miejscu gdzie powienine być link.'
-    ));
-    $wp_customize->add_setting('email', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('email', array(
-        'type'=>'input',
-        'priority'=>10,
-        'section'=>'event',
-        'label'=>'Adres e-mail dla powiadomień',
-        'description'=>'Wprowadź adres e-maila, na który zostanie wysłane powiadomienie.'
-    ));
-    $wp_customize->add_setting('title', array(
-        'type' => 'theme_mod',
-        'capability' => 'edit_theme_options',
-    ));
-    $wp_customize->add_control('title', array(
-        'type'=>'input',
-        'priority'=>10,
-        'section'=>'event',
-        'label'=>'Tytuł maila z powiadomieniem',
-        'description'=>'Wprowadź tytuł maila który zostanie wysłany z powiadomieniem.'
-    ));
-}
+require_once('includes/settings.php');
 // register endpoint to REST for theme options
-add_action('rest_api_init', 'register_rest_options');
-function register_rest_options()
-{
-    register_rest_route('wp/v2', '/options', array(
-        'methods' => 'GET',
-        'callback' => 'get_options'
-    ));
-}
-function get_options(WP_REST_Request $request)
-{
-    $logos = [
-        'header_logo_1'=>get_theme_mod('header_logo_1',false),
-        'header_logo_2'=>get_theme_mod('header_logo_2',false),
-        'header_logo_3'=>get_theme_mod('header_logo_3',false)
-    ];
-    $logos = array_map(fn($val)=>$val ?wp_get_attachment_image_src($val,'medium')[0]: $val, $logos);
-    return (object) array(
-        'title' => get_bloginfo('name'),
-        'descriptions' => get_bloginfo('description'),
-        'mourning' => get_theme_mod('mourning', false),
-        'add_event' => get_theme_mod('add_event', false),
-        'domains' => get_theme_mod('domains', ''),
-        'logo' => wp_get_attachment_image_src(get_theme_mod('logo', ''), 'medium')[0],
-        'placeholder' => wp_get_attachment_image_src(get_theme_mod('placeholder', ''), 'medium')[0],
-        'favicon'=>[
-            '512'=>get_site_icon_url(),
-            '128'=>get_site_icon_url(128),
-            '64'=>get_site_icon_url(64),
-            '32'=>get_site_icon_url(32)
-        ],
-        'header_logos'=>$logos 
-    );
-}
+require_once 'includes/api.php';
 // register endpoint to REST for events
 add_action('rest_api_init', 'register_rest_events');
 function register_rest_events()
