+++ packages/theme/includes/api.php
@@ -1,0 +1,3 @@
+<?php
+
+require_once 'api/options.php';
